load project dependencies depending on TFM
