+++ src/Compilers/Test/Core/DotNetCoreSdk.cs
@@ -22,7 +22,6 @@ namespace Roslyn.Test.Utilities
 
         static DotNetCoreSdk()
         {
-#if NETCOREAPP
             var dotNetExeName = "dotnet" + (Path.DirectorySeparatorChar == '/' ? "" : ".exe");
 
             bool DotNetExeExists(string? directory)
@@ -41,7 +40,6 @@ namespace Roslyn.Test.Utilities
             {
                 ExePath = Path.Combine(dotNetInstallDir, dotNetExeName);
             }
-#endif
         }
     }
 }
